test (ontology): create ontology mock utils
